+++ src/controllers/CartController.php
@@ -16,11 +16,13 @@ use craft\commerce\Plugin;
 use craft\elements\User;
 use craft\errors\ElementNotFoundException;
 use craft\helpers\Html;
+use craft\helpers\StringHelper;
 use craft\helpers\UrlHelper;
 use Throwable;
 use yii\base\Exception;
 use yii\base\InvalidConfigException;
 use yii\web\BadRequestHttpException;
+use yii\web\HttpException;
 use yii\web\NotFoundHttpException;
 use yii\web\Response;
 
@@ -86,6 +88,11 @@ class CartController extends BaseFrontEndController
         // When we are about to update the cart, we consider it a real cart at this point, and want to actually create it in the DB.
         $this->_cart = $this->_getCart(true);
 
+        // Can clear notices when updating the cart
+        if (($clearNotices = $this->request->getParam('clearNotices')) !== null) {
+            $this->_cart->clearNotices();
+        }
+
         // Set the custom fields submitted
         $this->_cart->setFieldValuesFromRequest('fields');
 
@@ -125,7 +132,7 @@ class CartController extends BaseFrontEndController
                 $purchasable['id'] = $purchasableId;
                 $purchasable['options'] = $options;
                 $purchasable['note'] = $note;
-                $purchasable['qty'] = $qty;
+                $purchasable['qty'] = (int) $qty;
 
                 $key = $purchasableId . '-' . LineItemHelper::generateOptionsSignature($options);
                 if (isset($purchasablesByKey[$key])) {
@@ -160,10 +167,9 @@ class CartController extends BaseFrontEndController
         // Update multiple line items in the cart
         if ($lineItems = $this->request->getParam('lineItems')) {
             foreach ($lineItems as $key => $lineItem) {
-
                 $lineItem = $this->_getCartLineItemById($key);
                 if ($lineItem) {
-                    $lineItem->qty = $this->request->getParam("lineItems.{$key}.qty", $lineItem->qty);
+                    $lineItem->qty = (int) $this->request->getParam("lineItems.{$key}.qty", $lineItem->qty);
                     $lineItem->note = $note = $this->request->getParam("lineItems.{$key}.note", $lineItem->note);
                     $lineItem->setOptions($this->request->getParam("lineItems.{$key}.options", $lineItem->getOptions()));
 
@@ -185,7 +191,7 @@ class CartController extends BaseFrontEndController
         }
 
         // Set if the customer should be registered on order completion
-        if ($registerUserOnOrderComplete = $this->request->getBodyParam('registerUserOnOrderComplete')) {
+        if ($this->request->getBodyParam('registerUserOnOrderComplete')) {
             $this->_cart->registerUserOnOrderComplete = true;
         }
 
@@ -235,7 +241,6 @@ class CartController extends BaseFrontEndController
 
     /**
      * @return Response|null
-     * @throws \craft\errors\MissingComponentException
      * @since 3.1
      */
     public function actionLoadCart()
@@ -269,33 +274,86 @@ class CartController extends BaseFrontEndController
 
         $cartCustomer = $cart->getCustomer();
 
-        if ($cartCustomer && $cartCustomer->userId && (!$this->_currentUser || $this->_currentUser->id != $cartCustomer->userId)) {
-            $error = Craft::t('commerce', 'You must be logged in to load this cart.');
-
-            if ($this->request->getAcceptsJson()) {
-                return $this->asErrorJson($error);
-            }
-
-            $this->setFailFlash($error);
-            return $this->request->getIsGet() ? $this->redirect($redirect) : null;
-        }
-
         $session = Craft::$app->getSession();
         $carts = Plugin::getInstance()->getCarts();
         $carts->forgetCart();
         $session->set($carts->getCartName(), $number);
-
-        $customers = Plugin::getInstance()->getCustomers();
-        $customers->forgetCustomer();
-        if ($cartCustomer) {
-            $session->set($customers::SESSION_CUSTOMER, $cartCustomer->id);
-        }
 
         if ($this->request->getAcceptsJson()) {
             return $this->asJson(['success' => true]);
         }
 
         return $this->request->getIsGet() ? $this->redirect($redirect) : $this->redirectToPostedUrl();
+    }
+
+    /**
+     * @return Response
+     * @since 3.3
+     */
+    public function actionComplete()
+    {
+        /** @var Plugin $plugin */
+        $plugin = Plugin::getInstance();
+        $this->requirePostRequest();
+
+        if (!$plugin->getSettings()->allowCheckoutWithoutPayment) {
+            throw new HttpException(401, Craft::t('commerce', 'You must make a payment to complete the order.'));
+        }
+
+        $this->_cart = $this->_getCart();
+        $errors = [];
+
+        // Check email address exists on order.
+        if (empty($this->_cart->email)) {
+            $errors['email'] = Craft::t('commerce', 'No customer email address exists on this cart.');
+        }
+
+        if ($plugin->getSettings()->allowEmptyCartOnCheckout && $this->_cart->getIsEmpty()) {
+            $errors['lineItems'] = Craft::t('commerce', 'Order can not be empty.');
+        }
+
+        if ($plugin->getSettings()->requireShippingMethodSelectionAtCheckout && !$this->_cart->getShippingMethod()) {
+            $errors['shippingMethodHandle'] = Craft::t('commerce', 'There is no shipping method selected for this order.');
+        }
+
+        if ($plugin->getSettings()->requireBillingAddressAtCheckout && !$this->_cart->billingAddressId) {
+            $errors['billingAddressId'] = Craft::t('commerce', 'Billing address required.');
+        }
+
+        if ($plugin->getSettings()->requireShippingAddressAtCheckout && !$this->_cart->shippingAddressId) {
+            $errors['shippingAddressId'] = Craft::t('commerce', 'Shipping address required.');
+        }
+
+        // Set if the customer should be registered on order completion
+        if ($this->request->getBodyParam('registerUserOnOrderComplete')) {
+            $this->_cart->registerUserOnOrderComplete = true;
+        }
+
+        if ($this->request->getBodyParam('registerUserOnOrderComplete') === 'false') {
+            $this->_cart->registerUserOnOrderComplete = false;
+        }
+
+        if (!empty($errors)) {
+            $this->_cart->addErrors($errors);
+        }
+
+
+        if (empty($errors)) {
+
+            $completedSuccess = false;
+
+            try {
+                $completedSuccess = $this->_cart->markAsComplete();
+            } catch (\Exception $exception) {
+                $completedSuccess = false;
+            }
+
+            if (!$completedSuccess) {
+                $this->_cart->addError('isComplete', Craft::t('commerce', 'Completing order failed.'));
+            }
+        }
+
+        return $this->_returnCart();
     }
 
     /**
@@ -342,7 +400,8 @@ class CartController extends BaseFrontEndController
 
         $updateCartSearchIndexes = Plugin::getInstance()->getSettings()->updateCartSearchIndexes;
 
-        if (!$this->_cart->validate($attributes) || !Craft::$app->getElements()->saveElement($this->_cart, false, false, $updateCartSearchIndexes)) {
+        // Do not clear errors, as errors could be added to the cart before _returnCart is called.
+        if (!$this->_cart->validate($attributes, false) || !Craft::$app->getElements()->saveElement($this->_cart, false, false, $updateCartSearchIndexes)) {
             $error = Craft::t('commerce', 'Unable to update cart.');
             $message = $this->request->getValidatedBodyParam('failMessage') ?? $error;
 
@@ -404,7 +463,11 @@ class CartController extends BaseFrontEndController
     {
         $cart = null;
 
-        if ($orderNumber = $this->request->getBodyParam('orderNumber')) {
+        // TODO Remove `orderNumber` param in 4.0
+        $orderNumber = $this->request->getBodyParam('orderNumber');
+        $orderNumber = $this->request->getBodyParam('number', $orderNumber);
+
+        if ($orderNumber) {
             // Get the cart from the order number
             $cart = Order::find()->number($orderNumber)->isCompleted(false)->one();