+++ src/controllers/OrdersController.php
@@ -11,6 +11,7 @@ use Craft;
 use craft\base\Element;
 use craft\base\Field;
 use craft\commerce\base\Gateway;
+use craft\commerce\base\Purchasable as PurchasableElement;
 use craft\commerce\base\PurchasableInterface;
 use craft\commerce\db\Table;
 use craft\commerce\elements\Order;
@@ -18,11 +19,14 @@ use craft\commerce\errors\CurrencyException;
 use craft\commerce\errors\RefundException;
 use craft\commerce\errors\TransactionException;
 use craft\commerce\gateways\MissingGateway;
+use craft\commerce\helpers\Currency;
+use craft\commerce\helpers\LineItem;
 use craft\commerce\helpers\Locale;
 use craft\commerce\helpers\Purchasable;
 use craft\commerce\models\Address;
 use craft\commerce\models\Customer;
 use craft\commerce\models\OrderAdjustment;
+use craft\commerce\models\OrderNotice;
 use craft\commerce\models\Transaction;
 use craft\commerce\Plugin;
 use craft\commerce\records\CustomerAddress;
@@ -131,7 +135,7 @@ class OrdersController extends Controller
      * @throws HttpException
      * @throws InvalidConfigException
      */
-    public function actionEditOrder($orderId, Order $order = null): Response
+    public function actionEditOrder($orderId, Order $order = null, $paymentForm = null): Response
     {
         $this->requirePermission('commerce-editOrders');
 
@@ -147,6 +151,7 @@ class OrdersController extends Controller
         }
 
         $variables['order'] = $order;
+        $variables['paymentForm'] = $paymentForm;
         $variables['orderId'] = $order->id;
 
         $transactions = $order->getTransactions();
@@ -412,13 +417,29 @@ class OrdersController extends Controller
             'billingAddress',
             'shippingAddress',
             'orderSite',
+            'notices',
+            'loadCartUrl',
         ];
+
+        $lineItems = $order->getLineItems();
+        $purchasableCpEditUrlByPurchasableId = [];
+        foreach ($lineItems as $lineItem) {
+            /** @var PurchasableElement $purchasable */
+            $purchasable = $lineItem->getPurchasable();
+            if (!$purchasable || isset($purchasableCpEditUrlByPurchasableId[$purchasable->id])) {
+                continue;
+            }
+
+            $purchasableCpEditUrlByPurchasableId[$purchasable->id] = $purchasable->getCpEditUrl();
+        }
 
         $orderArray = $order->toArray($orderFields, $extraFields);
 
+        // TODO merge this and the above line items loop into one.
         if (!empty($orderArray['lineItems'])) {
             foreach ($orderArray['lineItems'] as &$lineItem) {
                 $lineItem['showForm'] = ArrayHelper::isAssociative($lineItem['options']) || (is_array($lineItem['options']) && empty($lineItem['options']));
+                $lineItem['purchasableCpEditUrl'] = $purchasableCpEditUrlByPurchasableId[$lineItem['purchasableId']] ?? null;
             }
             unset($lineItem);
         }
@@ -718,6 +739,13 @@ class OrdersController extends Controller
         $order = $plugin->getOrders()->getOrderById($orderId);
         $gateways = $plugin->getGateways()->getAllGateways();
 
+        if ($paymentAmount = $request->getParam('paymentAmount')) {
+            $order->setPaymentAmount($paymentAmount);
+        }
+        if ($paymentCurrency = $request->getParam('paymentCurrency')) {
+            $order->setPaymentCurrency($paymentCurrency);
+        }
+
         $formHtml = '';
         /** @var Gateway $gateway */
         foreach ($gateways as $key => $gateway) {
@@ -881,6 +909,44 @@ class OrdersController extends Controller
         }
 
         return $this->redirectToPostedUrl();
+    }
+
+    /*
+     * @throws BadRequestHttpException
+     */
+    public function actionPaymentAmountData()
+    {
+        $this->requireAcceptsJson();
+        $this->requirePostRequest();
+        $paymentCurrencies = Plugin::getInstance()->getPaymentCurrencies();
+        $paymentCurrency = $this->request->getRequiredParam('paymentCurrency');
+        $paymentAmount = $this->request->getRequiredParam('paymentAmount');
+        $orderId = $this->request->getRequiredParam('orderId');
+        $order = Order::find()->id($orderId)->one();
+        $baseCurrency = $order->currency;
+
+        $baseCurrencyPaymentAmount = $paymentCurrencies->convertCurrency($paymentAmount, $paymentCurrency, $baseCurrency);
+        $baseCurrencyPaymentAmountAsCurrency = Craft::t('commerce', 'Pay {amount} of {currency} on the order.', ['amount' => Currency::formatAsCurrency($baseCurrencyPaymentAmount, $baseCurrency), 'currency' => $baseCurrency]);
+
+        $outstandingBalance = $order->outstandingBalance;
+        $outstandingBalanceAsCurrency = $order->outstandingBalanceAsCurrency;
+
+        $message = '';
+        if (Currency::round($baseCurrencyPaymentAmount) > Currency::round($outstandingBalance)) {
+            $baseCurrencyPaymentAmount = $outstandingBalance;
+            $baseCurrencyPaymentAmountAsCurrency = Craft::t('commerce', 'Pay {amount} of {currency} on the order.', ['amount' => $outstandingBalanceAsCurrency, 'currency' => $baseCurrency]);
+            $message = Craft::t('commerce', 'Order payment balance is {outstandingBalanceAsCurrency}. This is the maximum value that will be charged.', ['outstandingBalanceAsCurrency' => $outstandingBalanceAsCurrency]);
+        }
+
+        return $this->asJson([
+            'paymentCurrency' => $paymentCurrency,
+            'paymentAmount' => $paymentAmount,
+            'outstandingBalance' => $outstandingBalance,
+            'outstandingBalanceAsCurrency' => $outstandingBalanceAsCurrency,
+            'baseCurrencyPaymentAmountAsCurrency' => $baseCurrencyPaymentAmountAsCurrency,
+            'baseCurrencyPaymentAmount' => $baseCurrencyPaymentAmount,
+            'message' => $message
+        ]);
     }
 
     /**
@@ -1094,6 +1160,18 @@ class OrdersController extends Controller
         $order->message = $orderRequestData['order']['message'];
         $order->shippingMethodHandle = $orderRequestData['order']['shippingMethodHandle'];
 
+        $order->clearNotices();
+
+        // Create Notices on Order
+        $notices = [];
+        foreach ($orderRequestData['order']['notices'] as $notice) {
+            $notices[] = Craft::createObject([
+                'class' => OrderNotice::class,
+                'attributes' => $notice
+            ]);
+        }
+        $order->addNotices($notices);
+
         $dateOrdered = $orderRequestData['order']['dateOrdered'];
         if ($dateOrdered !== null) {
 
@@ -1146,7 +1224,7 @@ class OrdersController extends Controller
         $shippingAddress = null;
 
         // We need to create a new address if it belongs to a customer and the order is completed
-        if ($billingAddressId && $billingAddressId != 'new' &&  $order->isCompleted) {
+        if ($billingAddressId && $billingAddressId != 'new' && $order->isCompleted) {
             $belongsToCustomer = CustomerAddress::find()
                 ->where(['addressId' => $billingAddressId])
                 ->andWhere(['not', ['customerId' => null]])
@@ -1157,7 +1235,7 @@ class OrdersController extends Controller
             }
         }
 
-        if ($shippingAddressId && $shippingAddressId != 'new' &&  $order->isCompleted) {
+        if ($shippingAddressId && $shippingAddressId != 'new' && $order->isCompleted) {
             $belongsToCustomer = CustomerAddress::find()
                 ->where(['addressId' => $shippingAddressId])
                 ->andWhere(['not', ['customerId' => null]])
@@ -1217,12 +1295,13 @@ class OrdersController extends Controller
             $lineItemStatusId = $lineItemData['lineItemStatusId'];
             $options = $lineItemData['options'] ?? [];
             $qty = $lineItemData['qty'] ?? 1;
+            $uid = $lineItemData['uid'] ?? StringHelper::UUID();
 
             $lineItem = Plugin::getInstance()->getLineItems()->getLineItemById($lineItemId);
 
             if (!$lineItem) {
                 try {
-                    $lineItem = Plugin::getInstance()->getLineItems()->createLineItem($order->id, $purchasableId, $options, $qty, $note, $order);
+                    $lineItem = Plugin::getInstance()->getLineItems()->createLineItem($order->id, $purchasableId, $options, $qty, $note, $order, $uid);
                 } catch (\Exception $exception) {
                     $order->addError('lineItems', $exception->getMessage());
                     continue;
@@ -1235,16 +1314,13 @@ class OrdersController extends Controller
             $lineItem->privateNote = $privateNote;
             $lineItem->lineItemStatusId = $lineItemStatusId;
             $lineItem->setOptions($options);
+            $lineItem->uid = $uid;
 
             $lineItem->setOrder($order);
 
             // Deleted a purchasable while we had a purchasable ID in memory on the order edit page, unset it.
             if ($purchasableId && !Craft::$app->getElements()->getElementById($purchasableId)) {
                 $lineItem->purchasableId = null;
-            }
-
-            if ($order->getRecalculationMode() == Order::RECALCULATION_MODE_ALL) {
-                $lineItem->refreshFromPurchasable();
             }
 
             if ($order->getRecalculationMode() == Order::RECALCULATION_MODE_NONE) {
@@ -1414,12 +1490,14 @@ class OrdersController extends Controller
                 } else {
                     $row['priceAsCurrency'] = Craft::$app->getFormatter()->asCurrency($row['price'], $baseCurrency, [], [], true);
                 }
-                $row['isAvailable'] = $purchasable->getIsAvailable();
+                $row['isAvailable'] = Plugin::getInstance()->getPurchasables()->isPurchasableAvailable($purchasable);
                 $row['detail'] = [
                     'title' => Craft::t('commerce', 'Information'),
                     'content' => $purchasable->getSnapshot(),
                     'showAsList' => true,
                 ];
+                $row['newLineItemUid'] = StringHelper::UUID();
+                $row['newLineItemOptionsSignature'] = LineItem::generateOptionsSignature([]);
                 $purchasables[] = $row;
             }
         }