+++ src/elements/traits/OrderElementTrait.php
@@ -8,7 +8,8 @@
 namespace craft\commerce\elements\traits;
 
 use Craft;
-use craft\commerce\elements\actions\DownloadOrderPdf;
+use craft\commerce\elements\actions\CopyLoadCartUrl;
+use craft\commerce\elements\actions\DownloadOrderPdfAction;
 use craft\commerce\elements\actions\UpdateOrderStatus;
 use craft\commerce\elements\db\OrderQuery;
 use craft\commerce\elements\Order;
@@ -45,6 +46,16 @@ trait OrderElementTrait
     public function getFieldLayout()
     {
         return Craft::$app->getFields()->getLayoutByType(self::class);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function htmlAttributes(string $context): array
+    {
+        $attributes = parent::htmlAttributes($context);
+        $attributes['data-number'] = $this->number;
+        return $attributes;
     }
 
     /**
@@ -287,7 +298,7 @@ trait OrderElementTrait
             $sources[] = [
                 'key' => $key,
                 'status' => $orderStatus->color,
-                'label' => $orderStatus->name,
+                'label' => Craft::t('site', $orderStatus->name),
                 'criteria' => $criteriaStatus,
                 'defaultSort' => ['dateOrdered', 'desc'],
                 'badgeCount' => 0,
@@ -357,7 +368,7 @@ trait OrderElementTrait
             $elementService = Craft::$app->getElements();
 
             if (Plugin::getInstance()->getPdfs()->getHasEnabledPdf()) {
-                $actions[] = DownloadOrderPdf::class;
+                $actions[] = DownloadOrderPdfAction::class;
             }
 
             if (Craft::$app->getUser()->checkPermission('commerce-deleteOrders')) {
@@ -374,11 +385,17 @@ trait OrderElementTrait
             if (Craft::$app->getUser()->checkPermission('commerce-editOrders')) {
                 // Only allow mass updating order status when all selected are of the same status, and not carts.
                 $isStatus = strpos($source, 'orderStatus:');
-
-
                 if ($isStatus === 0) {
                     $updateOrderStatusAction = $elementService->createAction([
                         'type' => UpdateOrderStatus::class
+                    ]);
+                    $actions[] = $updateOrderStatusAction;
+                }
+
+                $isStatus = strpos($source, 'carts:');
+                if ($isStatus === 0) {
+                    $updateOrderStatusAction = $elementService->createAction([
+                        'type' => CopyLoadCartUrl::class
                     ]);
                     $actions[] = $updateOrderStatusAction;
                 }
@@ -469,6 +486,7 @@ trait OrderElementTrait
             $attributes[] = 'dateOrdered';
             $attributes[] = 'datePaid';
             $attributes[] = 'totalPaid';
+            $attributes[] = 'paidStatus';
             $attributes[] = 'totals';
         } else {
             $attributes[] = 'shortNumber';
@@ -592,7 +610,7 @@ trait OrderElementTrait
 
     /**
      * @param $miniTable Expects an array with rows of 'label', 'value' keys values.
-     * 
+     *
      * @return string
      */
     private function _miniTable($miniTable)