+++ src/migrations/m210113_093199_add_minMaxTotalType_to_shipping_rules.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace craft\commerce\migrations;
+
+use craft\db\Migration;
+use craft\db\Query;
+use Craft;
+
+/**
+ * m210113_093199_add_minMaxTotalType_to_shipping_rules migration.
+ */
+class m210113_093199_add_minMaxTotalType_to_shipping_rules extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function safeUp()
+    {
+        if (!$this->db->columnExists('{{%commerce_shippingrules}}', 'minMaxTotalType')) {
+            $this->addColumn(
+                '{{%commerce_shippingrules}}',
+                'minMaxTotalType',
+                $this->enum('minMaxTotalType', ['salePrice', 'salePriceWithDiscounts'])->notNull()->defaultValue('salePrice')
+            );
+        }
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown()
+    {
+        echo "m210113_093199_add_minMaxTotalType_to_shipping_rules cannot be reverted.\n";
+        return false;
+    }
+}