+++ src/models/LineItem.php
@@ -167,6 +167,11 @@ class LineItem extends Model
     public $dateUpdated;
 
     /**
+     * @var string UID
+     */
+    public $uid;
+
+    /**
      * @var PurchasableInterface Purchasable
      */
     private $_purchasable;
@@ -629,7 +634,7 @@ class LineItem extends Model
 
         /* @var $purchasable Purchasable */
         $purchasable = $this->getPurchasable();
-        if (!$purchasable || !$purchasable->getIsAvailable()) {
+        if (!$purchasable || !Plugin::getInstance()->getPurchasables()->isPurchasableAvailable($purchasable, $this->getOrder())) {
             return false;
         }
 
@@ -669,8 +674,8 @@ class LineItem extends Model
         $this->salePrice = Plugin::getInstance()->getSales()->getSalePriceForPurchasable($purchasable, $this->order);
         $this->taxCategoryId = $purchasable->getTaxCategoryId();
         $this->shippingCategoryId = $purchasable->getShippingCategoryId();
-        $this->sku = $purchasable->getSku();
-        $this->description = $purchasable->getDescription();
+        $this->setSku($purchasable->getSku());
+        $this->setDescription($purchasable->getDescription());
 
         // Check to see if there is a discount applied that ignores Sales for this line item
         $ignoreSales = false;
@@ -759,10 +764,7 @@ class LineItem extends Model
 
         foreach ($adjustments as $adjustment) {
             // Since the line item may not yet be saved and won't have an ID, we need to check the adjuster references this as it's line item.
-            $hasLineItemId = (bool)$adjustment->lineItemId;
-            $hasLineItem = (bool)$adjustment->getLineItem();
-
-            if (($hasLineItemId && $adjustment->lineItemId == $this->id) || ($hasLineItem && $adjustment->getLineItem() === $this)) {
+            if (($adjustment->lineItemId && $adjustment->lineItemId == $this->id) || (!$adjustment->lineItemId && $adjustment->getLineItem() === $this)) {
                 $lineItemAdjustments[] = $adjustment;
             }
         }