+++ src/web/assets/commercecp/src/js/CommerceVariantValuesInput.js
@@ -19,7 +19,6 @@ Craft.Commerce.VariantValuesInput = Craft.BaseInputGenerator.extend({
     updateTarget: function() {
         var sourceVal = this.$source.val();
         var targetVal = this.generateTargetValue(sourceVal);
-        console.log(sourceVal);
         this.$target.prop('checked', true);
     },
     onFormSubmit: function() {