+++ src/web/assets/commerceui/src/js/order/helpers/utils.js
@@ -23,6 +23,7 @@ export default {
                 lineItems: [],
                 orderAdjustments: [],
                 orderSiteId: draft.order.orderSiteId,
+                notices: draft.order.notices
             }
         }
 
@@ -53,6 +54,7 @@ export default {
             draftData.order.lineItems[lineItemKey].orderId = lineItem.orderId
             draftData.order.lineItems[lineItemKey].options = lineItem.options
             draftData.order.lineItems[lineItemKey].adjustments = []
+            draftData.order.lineItems[lineItemKey].uid = lineItem.uid
 
             lineItem.adjustments.forEach((adjustment, adjustmentKey) => {
                 draftData.order.lineItems[lineItemKey].adjustments[adjustmentKey] = {}