+++ src/web/assets/commerceui/src/js/order/store/index.js
@@ -19,7 +19,7 @@ export default new Vuex.Store({
         originalDraft: null,
         customers: [],
         orderData: null,
-        lastPurchasableIds: [],
+        recentlyAddedLineItems: [],
         unloadEventInit: false,
     },
 
@@ -385,6 +385,10 @@ export default new Vuex.Store({
 
                     throw errorMsg
                 });
+        },
+
+        clearRecentlyAddedLineItems({state}) {
+            state.recentlyAddedLineItems = []
         }
     },
 
@@ -433,8 +437,8 @@ export default new Vuex.Store({
             state.orderData = orderData
         },
 
-        updateLastPurchasableIds(state, lastPurchasableIds) {
-            state.lastPurchasableIds = lastPurchasableIds
+        updateRecentlyAddedLineItems(state, lineItemIdentifier) {
+            state.recentlyAddedLineItems.push(lineItemIdentifier)
         }
     }
 })