+++ tests/unit/models/AddressTest.php
@@ -32,7 +32,7 @@ class AddressTest extends Unit
      */
     public function testGetCpEditUrl() {
         $address = new Address(['id' => '1001']);
-        self::assertSame('http://craftcms.com/index.php?p=admin/commerce/addresses/1001', $address->getCpEditUrl());
+        self::assertSame('http://test.craftcms.test/index.php?p=admin/commerce/addresses/1001', $address->getCpEditUrl());
     }
 
     /**
@@ -358,8 +358,8 @@ class AddressTest extends Unit
     public function validateBusinessTaxIdDataProvider(): array
     {
         return [
-            ['123', false, [], false], // Don't validate
-            ['123', true, ['businessTaxId' => ['Invalid Business Tax ID.']], true], // validate - invalid
+            ['1123', false, [], false], // Don't validate
+            ['1123', true, ['businessTaxId' => ['Invalid Business Tax ID.']], true], // validate - invalid
             ['GB000472631', false, [], true], // validate - valid
             ['exists', false, [], true], // validate - valid - already exists
         ];