+++ tests/unit/services/OrdersTest.php
@@ -59,19 +59,19 @@ class OrdersTest extends Unit
 
     public function testGetOrderById()
     {
-        $order = $this->service->getOrderById($this->fixtureData['completed-new']['id']);
+        $order = $this->service->getOrderById($this->fixtureData->getElement('completed-new')->id);
 
         self::assertInstanceOf(Order::class, $order);
-        self::assertEquals($this->fixtureData['completed-new']['id'], $order->id);
+        self::assertEquals($this->fixtureData->getElement('completed-new')->id, $order->id);
     }
 
     public function testGetOrderByNumber()
     {
-        $order = $this->service->getOrderByNumber($this->fixtureData['completed-new']['number']);
+        $order = $this->service->getOrderByNumber($this->fixtureData->getElement('completed-new')->number);
 
         self::assertInstanceOf(Order::class, $order);
-        self::assertEquals($this->fixtureData['completed-new']['number'], $order->number);
-        self::assertEquals($this->fixtureData['completed-new']['id'], $order->id);
+        self::assertEquals($this->fixtureData->getElement('completed-new')->number, $order->number);
+        self::assertEquals($this->fixtureData->getElement('completed-new')->id, $order->id);
 
         $order = $this->service->getOrderByNumber('invalid');
 
@@ -80,23 +80,23 @@ class OrdersTest extends Unit
 
     public function testGetOrdersByCustomer()
     {
-        $orders = $this->service->getOrdersByCustomer($this->fixtureData['completed-new']['customerId']);
+        $orders = $this->service->getOrdersByCustomer($this->fixtureData->getElement('completed-new')->customerId);
 
         self::assertIsArray($orders);
-        self::assertCount(2, $orders);
+        self::assertCount(3, $orders);
         foreach ($orders as $order) {
-            self::assertTrue(in_array($order->id, [$this->fixtureData['completed-new']['id'], $this->fixtureData['completed-shipped']['id']]));
+            self::assertContains($order->id, [$this->fixtureData->getElement('completed-new')->id, $this->fixtureData->getElement('completed-new-past')->id, $this->fixtureData->getElement('completed-shipped')->id]);
         }
     }
 
     public function testGetOrdersByEmail()
     {
-        $orders = $this->service->getOrdersByEmail($this->fixtureData['completed-new']['email']);
+        $orders = $this->service->getOrdersByEmail($this->fixtureData->getElement('completed-new')->email);
 
         self::assertIsArray($orders);
-        self::assertCount(2, $orders);
+        self::assertCount(3, $orders);
         foreach ($orders as $order) {
-            self::assertEquals($this->fixtureData['completed-new']['email'], $order->email);
+            self::assertEquals($this->fixtureData->getElement('completed-new')->email, $order->email);
         }
     }
 }