+++ CHANGELOG.md
@@ -1,5 +1,96 @@
 # Release Notes for Craft Commerce
 
+## 3.3.3 - 2021-06-01
+
+### Added
+- Added the `productCount` and `variantCount` GraphQL queries. ([#1411](https://github.com/craftcms/commerce/issues/1411))
+
+### Changed
+- It’s now possible to sort products by their SKUs on product indexes. ([#2167](https://github.com/craftcms/commerce/issues/2167))
+- Products now have a `url` field when queried via GraphQL.
+
+### Fixed
+- Fixed a bug where it wasn’t possible to customize product search keywords via `EVENT_DEFINE_KEYWORDS`. ([#2142](https://github.com/craftcms/commerce/issues/2142))
+- Fixed a bug where the “Add Product to Sale” modal on Edit Product pages could be unresponsive when opened multiple times. ([#2146](https://github.com/craftcms/commerce/issues/2146))
+- Fixed an error that could occur if MySQL’s time zone tables weren’t populated yet. ([#2163](https://github.com/craftcms/commerce/issues/2163))
+- Fixed a PHP error that could occur when validating a product. ([#2138](https://github.com/craftcms/commerce/issues/2138))
+
+## 3.3.2 - 2021-05-18
+
+### Added
+- It’s now possible to create customer addresses in the control panel. ([#1324](https://github.com/craftcms/commerce/issues/1324))
+- Added `craft\commerce\events\PurchasableShippableEvent`.
+- Added `craft\commerce\services\Purchasables::EVENT_PURCHASABLE_SHIPPABLE`.
+- Added `craft\commerce\services\Purchasables::isPurchasableShippable()`.
+
+### Fixed
+- Customer search Ajax requests are now cancelled before sending new ones on Edit Order pages. ([#2137](https://github.com/craftcms/commerce/issues/2137))
+- Fixed an error that occurred when submitting a blank line item quantity from an Edit Order page, when running PHP 8. ([#2125](https://github.com/craftcms/commerce/issues/2125))
+- Fixed a bug where changes to addresses’ State fields on Edit Order pages weren’t persisting. ([#2136](https://github.com/craftcms/commerce/issues/2136))
+- Fixed a bug where charts weren’t always displaying the correct data for the date range, when running MySQL. ([#2117](https://github.com/craftcms/commerce/issues/2117))
+
+## 3.3.1.1 - 2021-05-09
+
+### Fixed
+- Fixed a bug that caused the billing address to be overridden by the shipping address on order completion. ([#2128](https://github.com/craftcms/commerce/issues/2128))
+
+## 3.3.1 - 2021-05-04
+
+### Added
+- Added `craft\commerce\events\RefundTransactionEvent::$refundTransaction`. ([#2081](https://github.com/craftcms/commerce/issues/2081))
+- Added `craft\commerce\services\Purchasables::EVENT_PURCHASABLE_AVAILABLE`.
+- Added `craft\commerce\services\Purchasables::isPurchasableAvailable()`.
+
+### Changed
+- Order condition formulas now include serialized custom field values. ([#2066](https://github.com/craftcms/commerce/issues/2066))
+- Replaced `date` to `datetime` filter of `orderHistory.dateCreated` attribute in status history tab in order edit page.
+
+### Fixed
+- Fixed a PHP error that occurred when changing a variant from having unlimited stock. ([#2111](https://github.com/craftcms/commerce/issues/2111))
+- Fixed a PHP error that occurred when passing the `registerUserOnOrderComplete` parameter to the `commerce/cart/complete` action.
+- Fixed a PHP error that occurred when attempting to retrieve an order notice that doesn’t exist. ([#2108](https://github.com/craftcms/commerce/issues/2108))
+- Fixed a bug where orders’ address IDs were `null` at the time `EVENT_AFTER_COMPLETE_ORDER` was triggered.
+- Fixed a bug where payment source error messages weren’t being returned correctly.
+
+## 3.3.0.1 - 2021-04-26
+
+### Fixed
+- Fixed a bug where an incorrect amount could be calculated when paying an outstanding balance in a non-primary currency.
+- Fixed a bug where shipping rules were enforcing the “Order Condition Formula” field as required. ([#2098](https://github.com/craftcms/commerce/issues/2098))
+- Fixed a bug where Base Discount and Per Item Discount fields could show negative values on the Edit Discount page. ([#2090](https://github.com/craftcms/commerce/issues/2090))
+
+## 3.3.0 - 2021-04-20
+
+### Added
+- Added support for partial payments. ([#585](https://github.com/craftcms/commerce/issues/585))
+- Carts can now display customer-facing notices on price changes and when items are automatically removed due to going out of stock. ([#2000](https://github.com/craftcms/commerce/pull/2000))
+- It’s now possible to set dynamic condition formulas on shipping rules. ([#1959](https://github.com/craftcms/commerce/issues/1959))
+- The Orders index page and Edit Order page now have a “Share cart” action, which generates a sharable URL that will load the cart into the user’s session, making it the active cart. ([#1386](https://github.com/craftcms/commerce/issues/1386))
+- Shipping rule conditions can now be based on an order’s discounted price, rather than its original price. ([#1948](https://github.com/craftcms/commerce/pull/1948))
+- Added the `allowCheckoutWithoutPayment` config setting.
+- Added the `allowPartialPaymentOnCheckout` config setting.
+- Added the `commerce/cart/complete` action.
+- Added `craft\commerce\base\GatewayInterface::supportsPartialPayment()`.
+- Added `craft\commerce\base\Gateway::supportsPartialPayment()`.
+- Added `craft\commerce\elements\Order::getLoadCartUrl()`.
+- Added `craft\commerce\services\Addresses::EVENT_BEFORE_PURGE_ADDRESSES`. ([#1627](https://github.com/craftcms/commerce/issues/1627))
+- Added `craft\commerce\services\PaymentCurrencies::convertCurrency()`.
+- Added `craft\commerce\test\fixtures\elements\ProductFixture::_getProductTypeIds()`.
+
+### Changed
+- Improved the line item editing workflow on the Edit Order page.
+- Line item descriptions now link to the purchasable’s edit page in the control panel. ([#2048](https://github.com/craftcms/commerce/issues/2048))
+- All front-end controllers now support passing the order number via a `number` param. ([#1970](https://github.com/craftcms/commerce/issues/1970))
+- Products are now resaved when a product type’s available tax or shipping categories change. ([#1933](https://github.com/craftcms/commerce/pull/1933))
+- Updated Dompdf to 1.0.2.
+
+### Deprecated
+- Deprecated `craft\commerce\services\Gateways::getGatewayOverrides()` and the `commerce-gateways.php` config file. Gateway-specific config files should be used instead. ([#1963](https://github.com/craftcms/commerce/issues/1963))
+
+### Fixed
+- Fixed a PHP 8 compatibility bug. ([#1987](https://github.com/craftcms/commerce/issues/1987))
+- Fixed an error that occurred when passing an unsupported payment currency to `craft\commerce\services\PaymentCurrencies::convert()`.
+
 ## 3.2.17.4 - 2021-04-06
 
 ### Fixed